💥 feat: convert to factory pattern
